refactor sys events
